ARGO-330:  fix and refactor version opening
include DorObjectHelper so that can_open_version? method is available.  switch the prepare method to use that check for determining whether an item can be opened for modification.

refactor prepare and open_version methods so that all operations performed on the object being opened are centralized inside the dor-services open_new_version method (which is being called anyway).

fix unit tests so that they test the new approach properly.  fix unit tests so that the existing @pid instance var gets used consistently instead of the inline string its equivalent to.