ARGO-367:  upgrade to blacklight 5:  start with 3 to 4 migration
a first step towards the goal of ultimately getting to blacklight 5: do the migration from 3 to 4, as per
https://github.com/projectblacklight/blacklight/releases/tag/v4.0.0.  this involves upgrading/adding/removing
some gems, fixing a template naming convention that's changed (both moving files and fixing references to those
templates in code), and fixing some stylesheet references.  note: blacklight-hierarchy has been disabled because
it's not compatible with blacklight 4, and i had trouble relaxing the dependency in my local gemspec.  i'm
almost certain it'll be necessary to add it back in for things to work, but i'm not sure when that will happen
since there's a messy set of dependency issues with blacklight-hierarchy and dor-services (blacklight-hierarchy
wants rails 4 and blacklight 5, dor-services wants rails 3).

this commit is too large, has some commented out work in progress, disables a key component, and is generally in
a very broken state.  but it's a first step in upgrading that's cohesive, and it's working just enough to be worth
saving.