add phantomjs to the Gemfile, so that installation of phantomjs is handled automatically
the addition of this gem allows the user to bundle and run rspec as they normally would, without the need to already have phantomjs installed (now that we use phantomjs for testing).  if phantomjs is already available on the system, nothing is downloaded or installed.

Gemfile was modified as per:  https://github.com/colszowka/phantomjs-gem#usage-with-poltergeistcapybara

spec_helper.rb was modified as per:  https://github.com/colszowka/phantomjs-gem#manual-setup

further info on the gem:  https://github.com/colszowka/phantomjs-gem