remove the argo:htaccess rake task
remove the code for the argo:htaccess rake task, as well as its invocation after restart in deploy.rb.

remove the default_htaccess_directives file that the rake task relied on, as well as the reference to that file in the :linked_files list in deploy.rb.

htaccess generation is no longer needed from argo.  we expect the htaccess file to change infrequently, and do not necessarily expect changes to coincide with web code deployments (which was the expectation set by the code being removed).  this will be managed as part of the deployment code by devops (currently using puppet).