Trade up some pending ModsulatorJobs tests
Add default values to some params.
Return from perform() when response is empty!
Use method block for ensure/rescue, etc.
Consolidate rescue blocks that are identical.
The rspec `integration` tag is now toast.

NOTE: perform spec was passing only because previous test left a copy of
the exactly-needed XML file as `smx.xml` and the glob for the check was
only as specific as `*.xml`! Tests should clean up after themselves and
all globs should be as specific as is practical to prevent collision.