Refactor param[:id] parsing using ApplicationController#find_druid
This provides more consistent application behavior.  It is unified
param[:id] parsing for Dor.find in ApplicationController#find_druid, so
that any expect on Dor.find in controllers can rely on it to get a
consistent DRUID input string pattern and more consistent exception
handling.  This also check that param[:id] is a valid DRUID and issues a
warning and log when it is not.