content type popup no longer displays layout
Because `content_type` is not a method name we are free to use in a
controller (becuase it's already used by Rails internals), I broke the
content_type metods out to their own controller.  This also serves to
reduce the size of ItemsController, which currently is doing too much.

Fixes #1039
Ref #1035