Use dor-services-app to register new objects
Do this instead of using `Dor::RegistrationService` within the `dor-services`
gem, which creates a service seam and reduces the reach and impact of
`dor-services`.

Note that this makes a small change to an integration spec that used to test a
side-effect of `Dor::RegistrationService` that is no longer as easy to test with
the new service boundary. We may want to revisit this integration spec if we
want the additional assurance this test affords us, say, if dor-services-app is
dockerized and added as a dependency of the dev/test environments for Argo.

Fixes #1212