Replace local dor-services-app client with gem
Fixes #1249

As we move away from the dor-services model (lots of modules included all over the place, resulting in objects with hundreds of methods) to the dor-services-app model (including said modules in a single place which sits behind RESTful HTTP APIs), we will wind up implementing dor-services-app clients across our codebases, like we already see in Argo. Instead of that, we ought to use a single client library for accessing the dor-services-app API, starting with Argo.

Configure the URI in one place so that we do not need to touch all the spots in the codebase that use the client.