Refactor to avoid checking for a newer file
build_datastream has a bunch of logic to see if there is a descMetadata.xml file in stacks
that is newer than the datastream and only then calls build_descMetadata_datastream which is
what actually refreshes from the catalog.  This change avoids a bunch of that complexity.

Additionally, there is no need to mess around with the datastream content, becuase
build_descMetadata_datastream takes care of that.