Use background job-based endpoint for virtual object creation
And in so doing, use as an opportunity to refactor the super awful CreateVirtualObjectsJob. I extracted two small objects from the job which greatly reduces its responsibility and knowledge.

Connects to #1323

Note that we CANNOT merge and deploy this until https://github.com/sul-dlss/puppet/pull/4470 is merged and puppet has run on the affected boxes. cc: @tallenaz