Remove Faraday config patch from initializer
This was added [recently](https://github.com/sul-dlss/argo/pull/1593) when virtual-merge was a synchronous operation, requiring Argo to hold a connection to dor-services-app open for up to 8 minutes for medium-sized batches. With the [pull request to make virtual-merge asynchronous](https://github.com/sul-dlss/argo/pull/1659) close to being merged, we should remove this; it will no longer be necessary and make in fact mask connection-related issues.