Fix importing of tags
This was broken in the bulk registration PR that omitted the import tags use case. I do not love the spec coverage I added, but it covers the bug this is fixing and beginning to add coverage of the bulk action form which was not covered and allowed this bug to go unnoticed.