Prefer using dor-services-client over Fedora connections
Fixes #2396

This refactors the `query_dor` method in the DescMetadataDownloadJob by replacing the call to `Dor.find` and `descMetadata` with corresponding dor-services-client calls. Note that I am not sure we still need the retry behavior that was patched in here previously, so I was inclined to make the smaller change and not remove this code. I might easily be convinced to do that, though.