Decouple registrations controller from Fedora
Fixes #2499

This commit adds a new service object with code that is largely copy-pasted from the dor-services gem, and it allows Argo to parse rights labels out of XML blobs without needing the dor-services dependency. Why? The defaultObjectRights attribute within the AdminPolicy `administrative` subset currently returns an XML blob but it is not an instance of dor-services rights metadata datastream, thus that dor-services method is not accessible. Note that the new service class should be covered (testwise) by the registration controller which is why additional unit tests were not added.