Disable save button after submitting edit tags form
Fixes #2729

This provides users immediate feedback that clicking the button worked as expected and suggest to them that the page transition should happen momentarily. It also prevents double-submits. Note that I attempted to add test coverage to the edit_tags feature spec, but I cannot inspect the button element after clicking the button because the selenium driver says the reference is stale. Does not seem worth the time to futz with Capybara to test such a tiny change.