Hide empty descriptive elements in displayed Cocina JSON
Fixes #3178

This commit hides empty descriptive elements in Cocina JSON when displayed, both on the catalog show page and in the items show view.

This is accomplished by wrapping the Cocina::Models instance with a presenter object that collaborates with a new service class that knows how to deeply remove blank values from enumerables (hashes and arrays). We lean on the Rails built-in method, `#compact_blank`, to do the heavy lifting.