Remove pid file when starting up rails dev server
If you start the containers with `docker compose up` and then stop them
with ctrl-c or a `docker compose stop` the tmp/pids/server.pid in the
working directory (shared with the docker container) is not deleted.
This means when you try to start it back up again you get an error about
the server already running.

I suspect there's a better way to do this, but this small change will remove the
server.pid before starting up `rails server`.