bulk action creation: remove bulk action form values from params passed along location URL in redirect when bulk action is created.
long druid lists were causing response handling upstream from argo to throw a 500 error because of the possibly enormous redirect location in the headers.

e.g. the apache error logs (argo_error_ssl.log) would report 'Premature end of script headers: bulk_actions, referer' when the user received a 500 in this case (despite the app logs showing a successful 303 response to the POST).  note that HTTP 303 responses should have a location.  perhaps the location header in the 303 was not parseable by apache once a certain size was exceeded?