Render Cocina on show page without metadata
Fixes #3583

This commit adds a new option to the `CocinaHashPresenter` that allows it to render a Cocina structure without metadata, and in this case it means without the lock, created timestamp, and modified timestamp. This new option is then used on the catalog show page. As part of this work, convert the CocinaHashPresenter spec to use factories instead of hand-crafted Cocina model instances and add a flag to the factories to retain past test behavior.