Lazy-load cocina object JSON on show page
Fixes #3626

This commit removes the burden of loading the full Cocina data structure from the show page, doing so lazily, only if the user interacts with the Cocina dropdown. This allows the show page to render more quickly. The approach taken in this commit follows similar work on lazy-loading events and technical metadata.