Replace rubocop with standard
The `standard` gem manages Rubocop upgrades and maintains strong opinions on the Ruby style guide so that we do not have to. This is a spike to explore what it would look like to replace Rubocop with standard and begin to evaluate which is a better fit for our team.

> [standard] aims to save uou (and others!) time in the same three ways:
>
> 1. **No configuration.** The easiest way to enforce consistent style in your project. Just drop it in.
> 1. **Automatically format code.** Just run `standardrb --fix` and say goodbye to messy or inconsistent code.
> 1. **Catch style issues & programmer errors early.** Save precious code review time by eliminating back-and-forth between reviewer & contributor.
>
> No decisions to make. It just works." ...

[Read more about standard](https://github.com/testdouble/standard). [Watch the standard lightning talk](https://www.youtube.com/watch?v=uLyV5hOqGQ8)

Note that the way we have knitted standard and rubocop together here, we are essentially using the standard style guide without being coupled to its linter tool. Otherwise put: we can (and probably should*) continue running `rubocop` for linting (vs. `standardrb`) and `rubocop -[aA]` for auto-fixing (vs. `standardrb --fix`).

* "should" because CI---via our ruby-rails orb, which uses the CircleCI ruby orb, which supports Rubocop but not standard---will continue using rubocop tooling, as will most/many of our IDEs. Better to be consistent with that practice in local dev than be surprised later, I reckon?