Add `metadata_source_ssim` facet field behind feature flag
Fixes #3977

This commit adds a new facet field to the Blacklight catalog if and only if the multi-valued metadata sources feature flag is on. It also flips this flag on in dev & test, and allows the user to flip it on for a given request by adding `?multivalued=true` as a query param in the URL.

It's not clear that there are any specs that could be extended to cover the two small changes to `app/controllers/` and I'm not sure it's worth the effort to write new specs for this when we'll be extensively testing this via the browser given we're dealing with Blacklight internals.