Remove controller code to flip Folio flag on based on a param
This turns out not to work as we expected. Mutating a `Settings` value changes **global** state, so when the `folio=true` query param is added to a request, the setting is tweaked not only for that request but flips the setting across the entire site.🤦🏻‍ Andrew does not use this feature anyway. Once I hear from Arcadia, that will unblock this PR.