Disable metadata refresh on show page during ILS cutover
Fixes #4031

This commit disables the metadata refresh button on the show page when the ILS cutover flag is flipped on. Not specified in the issue, I also included a small change to the "Refresh" label to make it clear to users *why* metadata refresh is not enabled. I pondered whether we should also include similar prevention code in the ItemsController, but since this operation is a POST, I figured it was very unlikely one of our users was going to bypass the UI controls and try to hack in a metadata refresh. Check me on this.