lint: get rid of duplicate method definitions discoverd by new rubocop rule
based on my tests using examples in rails console, the later definition should win (regardless of whether that definition is a full inline method definition using the `def` keyword, an `attr_reader` declaration for an instance variable, or a delegation (rails' `delegate` construct) to an collaborator object.

and luckily, the later definition seemed like the correct thing to keep in all cases (because the later-in-the-code-file definition was also the one that was added most recently, and/or because the later-in-the-code-file definition was the one that seemed the most sensible to keep for functional reasons).  so, this commit should not (hopefully) result in any functional changes, since the same method should be called as before this change.