Pin blacklight-hierarchy to 6.6.0 to avoid a new exception in 6.7.x
See: https://app.honeybadger.io/projects/49894/faults/122116465

This is a new exception that has been occurring in prod and other envs since the last blacklight-hierarchy bump earlier this week. Since Argo is stuck on BL7 for now and we do not know that need the bleeding edge of the hierarchy gem, pin the dependency for now and documented why it is pinned.