Open virtual object constituents in full body of window
Fixes #4835

Links in turbo-frames try to open within the frame, and the data returned by the links in this case do not match the turbo-frame ID, so the content is effectively wiped out. Instead, tell the browser to open these links in the full window to defeat the turbo-intercepted behavior.